Relocate Mind Flay interrupt proc
